company_brochure/requirements.txt
beautifulsoup4
python-dotenv
requests
openai
ipython